First Commit for Divinoacademy  Divinoacademy is an online e-learning platform that will be designed by God's grace to suite all students all over the world.
It will have the following features
1. An interactive study library for students to learn any possible and available course with a high level of focus and interactivity.
2. An interactive Exam simulation section, for students to practice any exam of their choice or exam they are expected to practice and get the best from it
3. Study materials such as periodic table, calculators, maths formulars, e.t.c
4. A gamified ui and interactive environment (like that of Dulingo)
5. A well organised, interactive, developed and assistive community for students
6. E.T.C.

In these section/commit I have at least achieved developing a little bit of the frontend Homepage, all with my mobile phone. But will still be working on it when I am chance, as am still in school.

But by the grace of God I will be looking forward to finish up the project (Divinoacademy) and launch for usability before the end of these year as Jesus tarries, as it is one of my major aim for the year.

These is my commit message for now.
Happy Coding in Christ Jesus.

@divinoacademy
17 - Feb. - 2023

Marvelous Innocent.
08168096890
marsteck.netlify.app